+++ yaat/middleware/__init__.py
@@ -1,3 +1,4 @@
 from yaat.middleware.base import BaseMiddleware
 from yaat.middleware.cors import CORSMiddleware
 from yaat.middleware.exception import ExceptionMiddleware
+from yaat.middleware.lifespan import LifespanMiddleware