+++ yaat/middleware/cors.py
@@ -1,12 +1,12 @@
 import re
 import typing
 
+from yaat.components import Headers
 from yaat.constants import HTTP_METHODS
-from yaat.datatypes import Headers
 from yaat.middleware import BaseMiddleware
 from yaat.requests import Request
 from yaat.responses import Response, TextResponse
-from yaat.types import ASGIApp
+from yaat.typing import ASGIApp
 
 
 # https://developer.mozilla.org/en-US/docs/Glossary/CORS-safelisted_request_header
@@ -32,10 +32,11 @@ class CORSMiddleware(BaseMiddleware):
     ):
         # https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
 
+        super().__init__(app)
+
         if "*" in allow_methods:
             allow_methods = HTTP_METHODS
 
-        self.app = app
         self.allow_origins = allow_origins
         self.allow_methods = allow_methods
         self.allow_headers = [header.lower() for header in allow_headers]