+++ docs/releasenotes.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+### 0.1.3
+
+- Added `StreamResponse` for HTTP response streaming.
+- Added server startup and shutdown events.
+
 ### 0.1.2
 
 - Added CORS middleware.